requirements.txt
Markdown
python-slugify
Flask
jrnl
pyyaml
